refact(QuickAddEditableView): implementation without react-hook-form
