+++ apps/store/package.json
@@ -76,7 +76,6 @@
     "react-day-picker": "8.10.1",
     "react-device-detect": "2.2.3",
     "react-dom": "18.3.1",
-    "react-hook-form": "7.51.4",
     "react-i18next": "14.1.1",
     "react-wrap-balancer": "1.1.0",
     "server-only": "0.0.1",
